refactor: move makeCommand to own directory
