+++ packages/mk/src/index.ts
@@ -1,3 +1,3 @@
 export * from './eventHub';
 export * from './makeContainer';
-export * from './lib/makeCommand';
+export * from './makeCommand';
