+++ packages/mk/src/makeCommand/index.ts
@@ -1,0 +1,1 @@
+export * from './makeCommand';
