+++ packages/mk/src/makeCommand/makeCommand.spec.ts
