+++ packages/mk/src/makeCommand/makeCommand.ts
